Consolidate util and utils
I'm not sure why there were two separate files for both of these; looks
like they both got moved up to the same directory when the project
workspace was removed in 3ee3a8304442c046ea9050f78993bd4b925a4b3f.

I moved util into utils because I see no need for there two be two
files, and it is a bit confusing. If I'm overlooking something please
let me know.
